File Upload and Download with pdfkit

+++ controllers/admin.js
@@ -1,7 +1,8 @@
 const { validationResult } = require('express-validator');
 
+const fileHelper = require('../util/file');
+
 const Product = require('../models/product');
-// const { ValidationError } = require('sequelize');
 
 exports.getAddProduct = (req, res, next) => {
       res.render('admin/edit-product', {
@@ -16,9 +17,25 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
       const title = req.body.title;
-      const imageUrl = req.body.imageUrl;
+      const image = req.file;
       const price = req.body.price;
       const description = req.body.description;
+      if (!image) {
+            return res.status(422).render('admin/edit-product', {
+                  pageTitle: 'Add Product',
+                  path: '/admin/add-product',
+                  editing: false,
+                  hasError: true,
+                  product: {
+                        title: title,
+                        price: price,
+                        description: description
+                  },
+                  errorMessage: 'Attached file is not an image.',
+                  validationErrors: []
+            });
+      }
+
       const errors = validationResult(req);
       if (!errors.isEmpty()) {
             return res.status(422).render('admin/edit-product', {
@@ -28,7 +45,6 @@ exports.postAddProduct = (req, res, next) => {
                   hasError: true,
                   product: {
                         title: title,
-                        imageUrl: imageUrl,
                         price: price,
                         description: description
                   },
@@ -36,6 +52,8 @@ exports.postAddProduct = (req, res, next) => {
                   validationErrors: errors.array()
             });
       }
+
+      const imageUrl = image.path;
       
       const product = new Product({ 
             title: title, 
@@ -51,21 +69,6 @@ exports.postAddProduct = (req, res, next) => {
                   res.redirect('/admin/products');
             })
             .catch(err => {
-                  // return res.status(500).render('admin/edit-product', {
-                  //       pageTitle: 'Add Product',
-                  //       path: '/admin/add-product',
-                  //       editing: false,
-                  //       hasError: true,
-                  //       product: {
-                  //             title: title,
-                  //             imageUrl: imageUrl,
-                  //             price: price,
-                  //             description: description
-                  //       },
-                  //       errorMessage: 'Database operation failed, please try again.',
-                  //       validationErrors: []
-                  // });
-                  // res.redirect('/500');
                   const error = new Error(err);
                   error.httpStatusCode = 500;
                   return next(error);
@@ -104,7 +107,7 @@ exports.postEditProduct = (req, res, next) => {
       const prodId = req.body.productId;
       const updatedTitle = req.body.title;
       const updatedPrice = req.body.price;
-      const updatedImageUrl = req.body.imageUrl;
+      const image = req.file;
       const updatedDesc = req.body.description;
       const errors = validationResult(req);
       if (!errors.isEmpty()) {
@@ -115,7 +118,6 @@ exports.postEditProduct = (req, res, next) => {
                   hasError: true,
                   product: {
                         title: updatedTitle,
-                        imageUrl: updatedImageUrl,
                         price: updatedPrice,
                         description: updatedDesc,
                         _id: prodId
@@ -132,7 +134,10 @@ exports.postEditProduct = (req, res, next) => {
             }
             product.title = updatedTitle;
             product.price = updatedPrice;
-            product.imageUrl = updatedImageUrl;
+            if (image) {
+                  fileHelper.deleteFile(product.imageUrl);
+                  product.imageUrl = image.path;
+            }
             product.description = updatedDesc;
             return product.save()
             .then(result => {
@@ -167,8 +172,15 @@ exports.getProductList = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
       const prodId = req.body.productId;
+      Product.findById(prodId)
+      .then(product => {
+            if (!product) {
+                  return next(new Error('Product Not Found!'));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({_id: prodId, userId: req.user._id});
+      })
       // Product.findByIdAndDelete(prodId)
-      Product.deleteOne({_id: prodId, userId: req.user._id})
       .then(()=> {
             console.log('Produtc Deleted Successfully!');
             res.redirect('/admin/products');
